// front/src/app/services/models/item-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ItemDto {
  cartId?: number;
  category?: string;
  code?: string;
  createdAt?: number;
  description?: string;
  image?: string;
  internalReference?: string;
  inventoryStatus?: 'INSTOCK' | 'LOWSTOCK' | 'OUTOFSTOCK';
  name?: string;
  oid?: number;
  price?: number;
  quantity?: number;
  rating?: number;
  shellId?: number;
  updatedAt?: number;
}